requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
asyncpg==0.29.0
twilio==8.10.0
pytesseract==0.3.10
Pillow==10.1.0
openai-whisper==20231117
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
schedule==1.2.0
redis==5.0.1
celery==5.3.4
